+++ src/Ch03/Tuples.hs
@@ -1,3 +1,4 @@
+{-# OPTIONS_GHC -Wno-unused-matches -Wno-incomplete-patterns #-}
 -- |
 -- Module      :  Ch03.Tuples
 -- Description :  Pattern Matching
@@ -12,8 +13,13 @@
 
 module Ch03.Tuples where
 
+-- added a GHC pragma to remove warnings -Wunused-matches -Wincomplete-patterns
+
 -- Further adventures, p. 52
 
+third :: (p, q, r) -> r
 third (a, b, c) = c
 
+-- added a GHC pragma to remove warnings -Wunused-matches -Wincomplete-patterns
+complicated :: (Bool, a, [b], Integer) -> (a, [b])
 complicated (True, a, x:xs, 5) = (a, xs)